+++ packages/app/src/systems/Account/services/account.test.ts
@@ -48,7 +48,7 @@ describe('AccountService', () => {
 
   it('should convert a map of accounts into an array', async () => {
     expect(
-      AccountService.fromMap({ [MOCK_ACCOUNT.address]: MOCK_ACCOUNT }),
+      AccountService.fromMap({ [MOCK_ACCOUNT.address]: MOCK_ACCOUNT })
     ).toEqual([MOCK_ACCOUNT]);
   });
 });