+++ packages/app/src/systems/Settings/machines/settingsMachine.ts
@@ -84,7 +84,7 @@ export const settingsMachine = createMachine(
         },
       }),
     },
-  },
+  }
 );
 
 export type SettingsMachine = typeof settingsMachine;