+++ packages/app/src/systems/SignUp/components/MnemonicRead/MnemonicRead.test.tsx
@@ -45,7 +45,7 @@ describe('MnemonicRead', () => {
             <SignUpProvider>{children}</SignUpProvider>
           </TestWrapper>
         ),
-      },
+      }
     );
     user = res.user;
   });