+++ packages/e2e-contract-tests/playwright/e2e/ForwardHalfCustomAsset.test.ts
@@ -14,6 +14,8 @@ import '../../load.envs';
 import { calculateAssetId, shortAddress } from '../../src/utils';
 import { testSetup } from '../utils';
 
+import { checkFee } from './utils';
+
 const { VITE_CONTRACT_ID } = process.env;
 
 test.describe('Forward Half Custom Asset', () => {
@@ -31,7 +33,7 @@ test.describe('Forward Half Custom Asset', () => {
     // Mint custom asset to wallet
     const contract = CustomAssetAbi__factory.connect(
       VITE_CONTRACT_ID!,
-      fuelWallet,
+      fuelWallet
     );
     const recipient: IdentityInput = {
       Address: {
@@ -53,7 +55,7 @@ test.describe('Forward Half Custom Asset', () => {
 
     const forwardHalfCustomAssetButton = getButtonByText(
       page,
-      'Forward Half Custom Asset',
+      'Forward Half Custom Asset'
     );
     await forwardHalfCustomAssetButton.click();
 
@@ -76,6 +78,7 @@ test.describe('Forward Half Custom Asset', () => {
 
     // test gas fee is correct
     await hasText(walletPage, 'Fee (network)');
-    await hasText(walletPage, '0.000000001 ETH');
+    const fee = bn.parseUnits('0.000000165');
+    await checkFee(walletPage, { minFee: fee.sub(100), maxFee: fee.add(100) });
   });
 });