+++ packages/sdk/src/Fuel.ts
@@ -21,7 +21,7 @@ export class Fuel extends FuelWalletConnection {
     createAddress: (address: string) => {
       // eslint-disable-next-line no-console
       console.warn(
-        'Do not use this method! It will be removed in the next release.',
+        'Do not use this method! It will be removed in the next release.'
       );
       return Address.fromString(address);
     },
@@ -48,14 +48,14 @@ export class Fuel extends FuelWalletConnection {
     // Listen for network changes and connect the provider
     // selected network from the user
     this.on(FuelWalletEvents.network, async (network) => {
-      FuelWeb3Privates.provider?.switchUrl(network.url);
+      FuelWeb3Privates.provider?.connect(network.url);
     });
 
     return FuelWeb3Privates.provider;
   }
 
   async getWallet(
-    address: string | AbstractAddress,
+    address: string | AbstractAddress
   ): Promise<FuelWalletLocked> {
     const provider = await this.getProvider();
     return new FuelWalletLocked(address, provider);
@@ -69,10 +69,10 @@ declare global {
   interface Document {
     addEventListener<K extends keyof FuelDocumentEvents>(
       type: K,
-      listener: (this: Document, ev: FuelDocumentEvents[K]) => void,
+      listener: (this: Document, ev: FuelDocumentEvents[K]) => void
     ): void;
     dispatchEvent<K extends keyof FuelDocumentEvents>(
-      ev: FuelDocumentEvents[K],
+      ev: FuelDocumentEvents[K]
     ): void;
   }
   interface Window {