+++ packages/test-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @fuel-wallet/test-utils
 
+## 0.13.4
+
+### Patch Changes
+
+- [#923](https://github.com/FuelLabs/fuels-wallet/pull/923) [`f683bae`](https://github.com/FuelLabs/fuels-wallet/commit/f683baeb6efbcc75561ac53c9c0d9d05f3bbae29) Thanks [@matt-user](https://github.com/matt-user)! - chore: bump fuels version
+
 ## 0.13.3
 
 ## 0.13.2